New outputs for separate calculation of ITG/TEM/ETG fluxes
For the neural network modelling, we may be interested in separating the contributions
from ITG, TEM, and ETG modes. This is to maintain clear threshold behaviour and
reduce complexity in the mappings, in case it is difficult to fit the total fluxes.
In that case, the NN will fit each flux contribution separately.

ETG fluxes are trivially separated due to different scale. However, including
self consistently ITG+TEM in the saturation rule leads to different results compared
to ITG and TEM separate superposition. In minor test cases so far, the differences are
only ~20% and tolerable. However, detailed testing is desired.

The choice whether to calculate and output the separate ITG, TEM, and ETG fluxes alongside
the total flux, is set by the new separateflux input variable.