Replaced SLATEC elliptical integrals by Fukusima
Solving the elliptical integrals for trapped particals uses up a large
part of the CPU time. This commit replaces the original SLATEC solvers
by the fast integral solvers by Toshio Fukushima. The results of the two
solvers agree within a few times the machine precision on the tested
interval (0..1) and are 30 to 50 times faster. They are based on the
following resources:

- [xcei.txt] [https://www.researchgate.net/publication/295857819_xceitxt_F90_package_of_complete_elliptic_integral_computation]
- [xelbd.txt] [https://www.researchgate.net/publication/303680876_xelbdtxt_Fortran_90_package_to_compute_general_completeincomplete_elliptic_integral_of_firstsecond_kind_in_singledouble_precision]